requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlmodel==0.0.16
asyncpg==0.29.0
psycopg[binary]
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt==4.0.1